refactor: move audit log handler builder to top-level
